requirements.txt
paho-mqtt
easydict
configparser
flask
pybluez
